scripts/requirements.txt
fastapi==0.103.1
ipython==8.15.0
pyserial==3.5
pytest==7.4.2
uvicorn==0.23.2